requirements.txt
aiofiles
aiohttp
alembic
anyio
bs4
bcrypt
dlib
deepface
fastapi
httpx
gtts
gradio
jinja2
librosa
matplotlib
numpy
opencv-python
python-dotenv
pytest
pytest-asyncio
pytest-asyncio
pytest-cov
pytest-tornasync
pytest-trio
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib
psutil
PyYAML
pyttsx3
requests
redis
sqlalchemy
spacy
scikit-learn
scipy
soundfile
torch
transformers
torchvision==0.20.1
tensorflow-cpu
tf-keras
textblob
uvicorn
whisper-openai
web.py
